requirements.txt
# Core dependencies
openai-agents>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# NLP dependencies
spacy>=3.7.0
dateparser>=1.2.0

# API integrations
todoist-api-python>=2.1.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# CLI dependencies
rich>=13.7.0
prompt-toolkit>=3.0.0
click>=8.1.0

# Utilities
aiohttp>=3.9.0
asyncio>=3.4.3
pytz>=2024.1